+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next version
 
+* Renamed `ftp-enforce-parens-in-return-single-element-tuple` to `ftp-disallow-parens-in-return-single-element-tuple`
+
 ## 25.3.25.1
 
 * Initial version